Refactor price parameters to use decimal type
Updated `FilterByPrice` and `FilterProductByPriceAsync` methods to use `decimal` instead of `double` for price parameters, enhancing precision. Commented out the `AddCartItem` method in `CartService.cs` as it is no longer needed.

+++ VTT_SHOP_API/Controllers/ProductsController.cs
@@ -76,7 +76,7 @@ namespace VTT_SHOP_API.Controllers
         }
 
         [HttpGet("filter-by-price")]
-        public async Task<IActionResult> FilterByPrice([FromQuery] double priceMin, [FromQuery] double priceMax)
+        public async Task<IActionResult> FilterByPrice([FromQuery] decimal priceMin, [FromQuery] decimal priceMax)
         {
             var result = await _product.FilterProductByPriceAsync(priceMin, priceMax);
             if (result.IsSuccess)
